tests: include: move aws client mock
Move aws_client_mock to the tests/include folder, such that mock can be
used by multiple unit tests.
It is a prerequisite for filter_aws ec2 tags unit tests.

It also formats a few files which had inconsistent style or leading
spaces.
