+++ tests/include/aws_client_mock.h
